Issue #1404: Fixes and refactoring in accordance with comments
