+++ api/src/test/java/com/epam/pipeline/test/creator/CommonCreatorConstants.java
@@ -20,8 +20,11 @@ import com.epam.pipeline.controller.Result;
 import com.fasterxml.jackson.core.type.TypeReference;
 
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 public final class CommonCreatorConstants {
 
@@ -29,15 +32,19 @@ public final class CommonCreatorConstants {
     public static final long ID_2 = 2L;
     public static final int TEST_INT = 4;
     public static final String TEST_STRING = "TEST";
-    public static final String TEST_STRING_2 = "TEST_2";
     public static final List<String> TEST_STRING_LIST = Collections.singletonList(TEST_STRING);
+    public static final byte[] TEST_ARRAY = {1, 1, 1};
+    public static final Map<String, String> TEST_STRING_MAP = Collections.singletonMap(TEST_STRING, TEST_STRING);
+    public static final Set<String> TEST_STRING_SET = Collections.emptySet();
+
     public static final TypeReference<Result<Integer>> INTEGER_TYPE = new TypeReference<Result<Integer>>() { };
     public static final TypeReference<Result<Object>> OBJECT_TYPE = new TypeReference<Result<Object>>() { };
     public static final TypeReference<Result<String>> STRING_TYPE = new TypeReference<Result<String>>() { };
     public static final TypeReference<Result<Map<String, String>>> STRING_STRING_MAP_TYPE =
             new TypeReference<Result<Map<String, String>>>() { };
-
     private CommonCreatorConstants() {
 
     }
+
+
 }
